Аватары, выбор валюты

+++ app/views.py
@@ -1,8 +1,11 @@
-from flask import render_template, abort, request, redirect, url_for, g, session, flash, jsonify
+from flask import render_template, abort, request, redirect, url_for, g, session, flash, jsonify, send_from_directory
 from flask_login import LoginManager, login_user, logout_user, current_user, login_required
-from app import app, forms, models, db
 from sqlalchemy import or_
 from datetime import datetime
+from werkzeug.utils import secure_filename
+from random import randint
+import os
+from app import app, forms, models, db, avatars
 from app.decorators import login_not_required
 
 lm = LoginManager()
@@ -44,7 +47,7 @@ def index():
         user = g.user,
         logs = logs)
 
-@app.route('/log/get')
+@app.route('/log/get', methods=['POST'])
 def log_get():
     logs_count = models.MoneyLog.query.filter(models.MoneyLog.user_id == g.user.id).count()
     logs = models.MoneyLog.query.filter(models.MoneyLog.user_id == g.user.id).order_by(models.MoneyLog.timestamp.desc(), models.MoneyLog.id.desc()).all()
@@ -72,7 +75,6 @@ def logout():
 @app.route('/auth', methods=['GET', 'POST'])
 @login_not_required
 def auth():
-#    if(g.user is not None and g.user.is_authenticated): return redirect(url_for('index'))
     form = forms.LoginForm()
     if(form.validate_on_submit() == True):
         cur_user = models.User.query.filter(or_(models.User.username == form.username.data, models.User.email == form.username.data)).first()
@@ -90,7 +92,6 @@ def auth():
 @app.route('/register', methods=['GET', 'POST'])
 @login_not_required
 def register():
-#    if(g.user is not None and g.user.is_authenticated): return redirect(url_for('index'))
     form = forms.RegisterForm()
     if(form.validate_on_submit() == True):
         if(models.User.query.filter(or_(models.User.username == form.username.data, models.User.email == form.email.data)).first() is not None):
@@ -132,14 +133,17 @@ def log_add():
         form = form,
         user = g.user)
 
-@app.route('/log/delete/<int:id_>', methods=['GET'])
+@app.route('/log/delete', methods=['POST'])
 @login_required
-def log_del(id_):
+def log_del():
+    id_ = request.form.get('id')
     d = models.MoneyLog.query.filter(models.MoneyLog.user_id == g.user.id, models.MoneyLog.id == id_).first_or_404()
-    db.session.delete(d)
-    db.session.commit()
-    flash("Запись успешно удалена", "success")
-    return redirect(url_for('index'))
+    if(d):
+        db.session.delete(d)
+        db.session.commit()
+        return {"status":"success"}
+    else:
+        return {"status":"fail", "description":"Запись не найдена"}
 
 @app.route('/log/edit/<int:id_>', methods=['GET' ,'POST'])
 @login_required
@@ -165,6 +169,11 @@ def log_edit(id_):
         form = form,
         user = g.user)
 
+@app.route('/profile/avatar/<path:filename>')
+@login_required
+def get_avatar(filename):
+    return send_from_directory(app.config['AVATARS_SAVE_PATH'], filename)
+
 @app.route('/profile/balance', methods=['POST'])
 @login_required
 def user_balance_edit():
@@ -180,3 +189,100 @@ def user_balance_edit():
     db.session.add(add)
     db.session.commit()
     return {"status": "success"}, 200
+
+@app.route('/profile/avatar/crop', methods=['GET', 'POST'])
+@login_required
+def avatar_crop():
+    form = forms.CropAvatarForm()
+    if(session.get('avatar_') is None):
+        return redirect(url_for('profile'))
+    if(form.validate_on_submit() == True):
+        x = request.form.get('x')
+        y = request.form.get('y')
+        w = request.form.get('w')
+        h = request.form.get('h')
+        filenames = avatars.crop_avatar(session['avatar_'], x, y, w, h)
+        os.remove(app.config['AVATARS_SAVE_PATH'] + "/" + session['avatar_'])
+        session.pop('avatar_')
+        if(g.user.avatar != None):
+            for n in ['s', 'm', 'l']:
+                a = g.user.avatar.split('_')
+                try:
+                    os.remove(app.config['AVATARS_SAVE_PATH'] + "/" + a[0] + "_" + n + a[1][1:])
+                except FileNotFoundError:
+                    pass
+        g.user.avatar = filenames[2]
+        db.session.commit()
+        flash('Аватар обновлён' ,"success")
+        return redirect(url_for('profile'))
+    return render_template('profile/avatar_crop.html',
+        user = g.user,
+        form = form,
+        title = "Обрезать аватар")
+
+@app.route('/profile', methods=['GET', 'POST'])
+@login_required
+def profile():
+    form = forms.UploadAvatarForm()
+    if(form.validate_on_submit() == True):
+        if('file' not in request.files):
+            form.file.errors.append("Некорректный файл")
+        file = request.files['file']
+        if(file.filename == ''):
+            form.file.errors.append("Не выбран файл")
+        elif(file.filename.split('.')[-1].lower() in app.config['ALLOWED_FILE_EXTENSIONS']):
+            session['avatar_'] = avatars.save_avatar(file)
+            return redirect(url_for('avatar_crop'))
+    return render_template("profile.html",
+        title = "Профиль",
+        user = g.user,
+        form = form)
+
+@app.route('/profile/changepass', methods=['GET', 'POST'])
+def profile_changepass():
+    form = forms.ProfileChangepass()
+    if(form.validate_on_submit() == True):
+        if(g.user.check_password(form.cur_password.data) != True):
+            form.cur_password.errors.append("Текущий пароль введён неверно")
+        elif(form.cur_password.data == form.new_password.data):
+            form.new_password.errors.append("Новый пароль должен отличаться от старого")
+        else:
+            g.user.set_password(form.new_password.data)
+            flash("Пароль успешно изменён", "success")
+            return redirect(url_for('index'))
+    return render_template("/profile/changepass.html",
+        title = "Смена пароля",
+        form = form,
+        user = g.user)
+
+@app.route('/profile/delete')
+@login_required
+def profile_delete():
+    d = models.MoneyLog.query.filter(models.MoneyLog.user_id == g.user.id).all()
+    for n in d:
+        db.session.delete(n)
+    if(g.user.avatar):
+        for n in ['s', 'm', 'l']:
+            a = g.user.avatar.split('_')
+            try:
+                os.remove(app.config['AVATARS_SAVE_PATH'] + "/" + a[0] + "_" + n + a[1][1:])
+            except FileNotFoundError:
+                pass
+    user = models.User.query.filter(models.User.id == g.user.id).first()
+    db.session.delete(user)
+    db.session.commit()
+    logout_user()
+    flash("Аккаунт успешно удалён" ,"success")
+    return redirect(url_for('auth'))
+
+@app.route('/profile/change_currency', methods=['POST'])
+def change_currency():
+    currency = request.form['currency']
+    if(currency not in g.user.currency_list):
+        flash("Валюты нет в списке", "error")
+        return {"status":"fail", "description":"Валюты нет в списке"}
+    flash("Валюта изменена", "success")
+    if(currency == g.user.currency): return {"status": "success"}
+    g.user.currency = currency
+    db.session.commit()
+    return {"status":"success"}
